+++ src/op.rs
@@ -7,13 +7,13 @@ use super::var_store::{VarIndex, VarStore};
 
 pub struct Operation {
     pub name: String,
-    pub num_inputs: u64,
-    pub num_outputs: u64,
+    pub num_inputs: usize,
+    pub num_outputs: usize,
     pub build: Box<Fn(&matrix::Context, &mut dl::Graph, &mut VarStore, &[Option<VarIndex>], &[VarIndex])>,
 }
 
 impl Operation {
-    pub fn new<F>(name: String, num_inputs: u64, num_outputs: u64, build: F) -> Self
+    pub fn new<F>(name: String, num_inputs: usize, num_outputs: usize, build: F) -> Self
         where F: Fn(&matrix::Context, &mut dl::Graph, &mut VarStore,
                     &[Option<VarIndex>], &[VarIndex]) + 'static {
         Operation {
